refactor: lander reorganized based on new super classes
