+++ src/main/java/dev/galacticraft/mod/content/entity/vehicle/AbstractLanderEntity.java
@@ -22,21 +22,124 @@
 
 package dev.galacticraft.mod.content.entity.vehicle;
 
+import dev.galacticraft.api.entity.IgnoreShift;
+import dev.galacticraft.api.rocket.entity.PlayerRideable;
+import dev.galacticraft.mod.screen.ParachestMenu;
+import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.HasCustomInventoryScreen;
+import net.minecraft.world.entity.MoverType;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
 
-public abstract class AbstractLanderEntity extends GCPlayerRideableVehicleEntity {
+public abstract class AbstractLanderEntity extends GCFueledVehicleEntity implements IgnoreShift, PlayerRideable, HasCustomInventoryScreen, ExtendedScreenHandlerFactory {
+
+    // **************************************** FIELDS ****************************************
+
+    // **************************************** CONSTRUCTOR ****************************************
+
     public AbstractLanderEntity(EntityType<?> entityType, Level level) {
         super(entityType, level);
+    }
+
+    // **************************************** DATA ****************************************
+
+    public Item getDropItem() {
+        return ItemStack.EMPTY.getItem();
     }
 
     public abstract boolean shouldMove();
 
     public abstract boolean shouldSpawnParticles();
 
-    public Item getDropItem() {
-        return ItemStack.EMPTY.getItem();
+    @Override
+    public boolean shouldIgnoreShiftExit() {
+        return !onGround();
     }
+
+    // **************************************** FUEL ****************************************
+
+    @Override
+    public int getFuelTankCapacity() {
+        return 100;
+    }
+
+    // **************************************** INTERACTION ****************************************
+
+    @Override
+    public float getDamageMultiplier() {
+        return 5;
+    }
+
+    @Override
+    public float getMaxDamage() {
+        return 100;
+    }
+
+    @Override
+    public InteractionResult interact(Player player, InteractionHand hand) {
+        if (this.level().isClientSide) {
+            if (!this.onGround()) {
+                return InteractionResult.FAIL;
+            }
+
+            if (!this.getPassengers().isEmpty()) {
+                this.ejectPassengers();
+            }
+
+            return InteractionResult.SUCCESS;
+        }
+
+        if (this.getPassengers().isEmpty()) {
+            openCustomInventoryScreen(player);
+            return InteractionResult.SUCCESS;
+        } else if (player instanceof ServerPlayer) {
+            if (!this.onGround()) {
+                return InteractionResult.FAIL;
+            }
+
+            this.ejectPassengers();
+            return InteractionResult.SUCCESS;
+        } else {
+            return InteractionResult.SUCCESS;
+        }
+    }
+
+    // **************************************** TICK ****************************************
+
+    @Override
+    public void move(MoverType movementType, Vec3 movement) {
+        if (shouldMove())
+            super.move(movementType, movement);
+    }
+
+    // **************************************** INVENTORY ****************************************
+
+    @Nullable
+    @Override
+    public AbstractContainerMenu createMenu(int containerId, Inventory inventory, Player player) {
+        return new ParachestMenu(containerId, inventory, this);
+    }
+
+    @Override
+    public void openCustomInventoryScreen(Player player) {
+        player.openMenu(this);
+    }
+
+    @Override
+    public void writeScreenOpeningData(ServerPlayer player, FriendlyByteBuf buf) {
+        buf.writeBoolean(false);
+        buf.writeVarInt(this.inventory.size());
+    }
+
 }
